Install hints
Since the extension from now on no longer requires a specific numpy version, but is compiled against the current numpy version and thus depends on it, the installation cannot be run in an isolated environment. The latter would install a different numpy version and thus render the extension unusable.
Therefore, the `--no-build-isolation` and `--no-deps` options must be used during installation with pip.